Add files via upload
Here is a Python program which demonstrates the no of volcanoes in USA and also shows the world population with the help of folium library of python
Folium is a Python library used for visualizing geospatial data. It is easy to use and yet a powerful library. Folium is a Python wrapper for Leaflet. js which is a leading open-source JavaScript library for plotting interactive maps.